requirements.txt
fastapi==0.104.1
pywa==1.22.1
openai==1.3.8
python-dotenv==1.0.0
uvicorn==0.24.0
httpx==0.25.2